Select format (webp/gif). Defaults to webp.

+++ src/endpoints/ttp.ts
@@ -14,10 +14,11 @@ const BLACK = '#000000'
 const WHITE = '#FFFFFF'
 const COLORS = ['#FFA07A', '#FFFFE0', '#98FB98', '#ADD8E6', '#F08080', '#EEE8AA', '#90EE90', '#87CEFA', '#FFB6C1', '#FFC0CB']
 
-export const ttp = (req: Request, res: Response) => {
+export const ttp = async (req: Request, res: Response) => {
     // Get text from request query
     const text = <string>req.query.text
     const lines = limitSplit(text)
+    const format = <'webp' | 'gif'>req.query.format || 'webp'
 
     let fontSize = FONT_SIZE
     if (lines.length > MAX_LINES) {
@@ -47,13 +48,21 @@ export const ttp = (req: Request, res: Response) => {
     })
     encoder.updateFrame()
     y = x - totalHeight / 2
-    res.type('image/gif').end(encoder.finish())
+
+    // Send preferred format
+    if (format === 'webp') {
+        const webp = await gifToWebp(encoder.finish())
+        res.type('image/webp').sendFile(webp, async () => await fs.unlink(webp))
+    } else {
+        res.type('image/gif').end(encoder.finish())
+    }
 }
 
 export const attp = async (req: Request, res: Response) => {
     // Get text from request query
     const text = <string>req.query.text
     const lines = limitSplit(text)
+    const format = <'webp' | 'gif'>req.query.format || 'webp'
 
     let fontSize = FONT_SIZE
     if (lines.length > MAX_LINES) {
@@ -84,6 +93,12 @@ export const attp = async (req: Request, res: Response) => {
         encoder.updateFrame()
         y = x - totalHeight / 2
     })
-    const webp = await gifToWebp(encoder.finish())
-    res.type('image/webp').sendFile(webp, async () => await fs.unlink(webp))
+
+    // Send preferred format
+    if (format === 'webp') {
+        const webp = await gifToWebp(encoder.finish())
+        res.type('image/webp').sendFile(webp, async () => await fs.unlink(webp))
+    } else {
+        res.type('image/gif').end(encoder.finish())
+    }
 }
